requirements.txt
flask
langchain
google-generativeai
langchain-google-genai
from_root
python-dotenv
-e .
